requirements.txt
Django==4.2.1
Django==4.2.2
djangorestframework==3.14.0
pyotp==2.8.0
